requirements.txt
flask
python-telegram-bot
scikit-learn
pandas
requests
beautifulsoup4
sqlite3
